lib: utils: Add support for RAS agent in OpenSBI
The RAS agent in OpenSBI get information about various error sources
in the system and their properties. The properties include the reri
bank addresse, the notification type, SSE vector, etc. It parses the
error sources and allocates buffers for APEI Error Status block for
each error source. On an error, the reri registers of the error sources
are read and their corresponding CPER record is generated and logged.

Signed-off-by: Himanshu Chauhan <hchauhan@ventanamicro.com>